refactor: split components into separate files
